requirements.txt
boto
troposphere
awacs
pyyaml